+++ apps/omsorgspengerutbetaling-arbeidstaker-soknad/src/storybook/stories/steps/Legeerklæring.stories.tsx
@@ -1,5 +1,4 @@
 import { Meta, StoryFn } from '@storybook/react-vite';
-import * as React from 'react';
 import LegeerklæringStep from '../../../app/søknad/steps/legeerklæring/LegeerklæringStep';
 import { withAmplitudeProvider } from '../../decorators/withAmplitudeProvider';
 import { withFormikWrapper } from '../../decorators/withFormikWrapper';