+++ apps/omsorgspengerutbetaling-soknad/src/storybook/stories/pages/KvitteringPage.stories.tsx
@@ -1,5 +1,4 @@
 import { Meta, StoryFn } from '@storybook/react-vite';
-import * as React from 'react';
 import KvitteringPage from '../../../app/pages/kvittering/KvitteringPage';
 import { withAmplitudeProvider } from '../../decorators/withAmplitudeProvider';
 import { withIntl } from '../../decorators/withIntl';