+++ apps/ungdomsytelse-veileder/CHANGELOG.md
@@ -1,5 +1,30 @@
 # @navikt/sif-demo-app
 
+## 0.5.30
+
+### Patch Changes
+
+- Pakkeoppdateringer - minor/patch + sentry 10
+- Updated dependencies
+    - @navikt/sif-common-core-ds@10.2.26
+    - @navikt/ung-common@0.0.20
+    - @navikt/sif-common-ui@0.10.2
+    - @navikt/sif-common-api@0.4.24
+
+## 0.5.29
+
+### Patch Changes
+
+- Pakkeoppdateringer - minor/patch
+- Updated dependencies
+    - @navikt/sif-common-formik-ds@2.4.22
+    - @navikt/sif-common-core-ds@10.2.25
+    - @navikt/sif-common-api@0.4.23
+    - @navikt/sif-common-env@0.1.5
+    - @navikt/sif-validation@1.0.3
+    - @navikt/sif-common-ui@0.10.1
+    - @navikt/ung-common@0.0.19
+
 ## 0.5.28
 
 ### Patch Changes